elf: Move Elf64AddrRange to separate module
In the effort of splitting up the ELF parser code, move the
Elf64AddrRange implementation to its own module.
